Unify impl of `EventLoop::run` for `run_return` platforms
