+++ src/platform_impl/macos/event_loop.rs
@@ -5,7 +5,7 @@ use std::{
     marker::PhantomData,
     mem,
     os::raw::c_void,
-    panic::{catch_unwind, resume_unwind, AssertUnwindSafe, RefUnwindSafe, UnwindSafe},
+    panic::{catch_unwind, resume_unwind, RefUnwindSafe, UnwindSafe},
     process, ptr,
     rc::{Rc, Weak},
     sync::mpsc,
@@ -176,29 +176,6 @@ impl<T> EventLoop<T> {
 
     pub fn window_target(&self) -> &RootWindowTarget<T> {
         &self.window_target
-    }
-
-    pub fn run<F>(mut self, callback: F) -> !
-    where
-        F: 'static + FnMut(Event<'_, T>, &'static RootWindowTarget<T>, &mut ControlFlow),
-    {
-        // SAFETY: `process::exit` will terminate the entire program before `self` is
-        // dropped, and `catch_unwind` prevents control from from exiting this function
-        // by panicking, therefore it will live for the rest of the program ('static).
-        //
-        // I believe this pointer casting is the correct way to do it because that's how
-        // `Box::leak` is implemented (https://doc.rust-lang.org/1.60.0/src/alloc/boxed.rs.html#1147-1152)
-        let this: &'static mut Self = unsafe { &mut *(&mut self as *mut Self) };
-        // Note: we don't touch `callback` again if this unwinds, so it doesn't matter
-        // if it's unwind safe.
-        let exit_code = match catch_unwind(AssertUnwindSafe(|| this.run_return(callback))) {
-            Ok(code) => code,
-            // 101 seems to be the status code Rust uses for panics.
-            // Note: the panic message gets printed before unwinding, so we don't have to print it
-            // ourselves.
-            Err(_) => 101,
-        };
-        process::exit(exit_code);
     }
 
     pub fn run_return<'a, F>(&'a mut self, callback: F) -> i32
